services/recommendations-service/requirements.txt
flask>=2.3.3
flask-cors>=4.0.0
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
requests>=2.31.0
setuptools>=68.0.0
gunicorn==21.2.0 